+++ package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "apollo-server": "^2.8.2",
     "bcrypt": "^3.0.6",
+    "dotenv": "^8.1.0",
     "graphql": "^14.4.2",
     "jsonwebtoken": "^8.5.1",
     "mysql2": "^1.6.5",
